src/app/shared/classes/methods.spec.ts
import { TestBed } from '@angular/core/testing';
import { Renderer2 } from '@angular/core';
import { Methods } from './methods'; // Import the class under test

// Mock the Renderer2 for more isolated testing
const mockRenderer2 = jasmine.createSpyObj('Renderer2', ['setStyle']);

describe('Methods', () => {
  let methods: Methods;

  beforeEach(async () => {
    await TestBed.configureTestingModule({
      providers: [
        { provide: Renderer2, useValue: mockRenderer2 }, // Inject the mock Renderer2
      ],
    }).compileComponents();
    methods = new Methods(); // Instantiate the class
  });

  // Test the pxToRem function
  it('should convert pixels to rem correctly', () => {
    expect(methods.openable.pxToRem(10)).toBe('0.625');
    expect(methods.openable.pxToRem(24)).toBe('1.5');
    expect(methods.openable.pxToRem(50)).toBe('3.125');
  });

  // Test the getDivElement function (considering different scenarios)
  it('should get the div element with elementId when provided', () => {
    const mockElement = createMockElement(); // Create a mock element
    const elementId = 'target-div';
    methods.openable.getDivElement({
      target: mockElement,
      bubbles: false,
      cancelBubble: false,
      cancelable: false,
      composed: false,
      currentTarget: null,
      defaultPrevented: false,
      eventPhase: 0,
      isTrusted: false,
      returnValue: false,
      srcElement: null,
      timeStamp: 0,
      type: '',
      composedPath: function (): EventTarget[] {
        throw new Error('Function not implemented.');
      },
      initEvent: function (type: string, bubbles?: boolean | undefined, cancelable?: boolean | undefined): void {
        throw new Error('Function not implemented.');
      },
      preventDefault: function (): void {
        throw new Error('Function not implemented.');
      },
      stopImmediatePropagation: function (): void {
        throw new Error('Function not implemented.');
      },
      stopPropagation: function (): void {
        throw new Error('Function not implemented.');
      },
      NONE: 0,
      CAPTURING_PHASE: 1,
      AT_TARGET: 2,
      BUBBLING_PHASE: 3
    }, elementId);
    expect(mockRenderer2.setStyle).toHaveBeenCalledWith(
      mockElement,
      'height',
      /auto/
    );
  });

  it('should get the parent div element when elementId is not provided', () => {
    const mockElement = createMockElement();
    const mockParent = createMockElement();
    mockElement.parentNode = mockParent;
    methods.openable.getDivElement({
      target: mockElement,
      bubbles: false,
      cancelBubble: false,
      cancelable: false,
      composed: false,
      currentTarget: null,
      defaultPrevented: false,
      eventPhase: 0,
      isTrusted: false,
      returnValue: false,
      srcElement: null,
      timeStamp: 0,
      type: '',
      composedPath: function (): EventTarget[] {
        throw new Error('Function not implemented.');
      },
      initEvent: function (type: string, bubbles?: boolean | undefined, cancelable?: boolean | undefined): void {
        throw new Error('Function not implemented.');
      },
      preventDefault: function (): void {
        throw new Error('Function not implemented.');
      },
      stopImmediatePropagation: function (): void {
        throw new Error('Function not implemented.');
      },
      stopPropagation: function (): void {
        throw new Error('Function not implemented.');
      },
      NONE: 0,
      CAPTURING_PHASE: 1,
      AT_TARGET: 2,
      BUBBLING_PHASE: 3
    });
    expect(mockRenderer2.setStyle).toHaveBeenCalledWith(
      mockParent,
      'height',
      /auto/
    );
  });

  // ... Add more tests for openDiv, openDivPlus, openDivTarget functions, covering various input scenarios and expected behaviors
});

// Helper function to create mock elements for testing
function createMockElement(styles = {}) {
  const mockElement = jasmine.createSpyObj('HTMLElement', [
    'querySelector',
    'dataset',
  ]);
  mockElement.style = styles;
  return mockElement;
}
